Wizard/ImageSourceSelect: move auth section below image source controls
