+++ src/Components/CreateImageWizard/steps/ImageOutput/components/ImageSourceSelect/RegistryAuthSection.tsx
